refactor: Rename auth.schemas to auth.models
