+++ tracecat/api/routers/users.py
@@ -6,7 +6,7 @@ from sqlmodel import select
 from sqlmodel.ext.asyncio.session import AsyncSession
 
 from tracecat.auth.credentials import authenticate_user_access_level
-from tracecat.auth.schemas import UserRead
+from tracecat.auth.models import UserRead
 from tracecat.db.engine import get_async_session
 from tracecat.db.schemas import User
 from tracecat.logger import logger
